Move Link configuration to payment element
Committed-By-Agent: codex
