+++ paymentsheet/src/main/java/com/stripe/android/checkout/LinkConfigurationMapper.kt
@@ -1,10 +1,11 @@
 package com.stripe.android.checkout
 
+import com.stripe.android.elements.PaymentElement
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentsheet.PaymentSheet
 
 @OptIn(CheckoutSessionPreview::class)
-internal fun CheckoutController.Configuration.LinkConfiguration.State.asPaymentSheet():
+internal fun PaymentElement.Configuration.LinkConfiguration.State.asPaymentSheet():
     PaymentSheet.LinkConfiguration =
     PaymentSheet.LinkConfiguration(
         display = display.asPaymentSheet(),
@@ -16,12 +17,12 @@ internal fun CheckoutController.Configuration.LinkConfiguration.State.asPaymentS
     )
 
 @OptIn(CheckoutSessionPreview::class)
-private fun CheckoutController.Configuration.LinkConfiguration.Display.asPaymentSheet():
+private fun PaymentElement.Configuration.LinkConfiguration.Display.asPaymentSheet():
     PaymentSheet.LinkConfiguration.Display = when (this) {
-    CheckoutController.Configuration.LinkConfiguration.Display.Automatic ->
+    PaymentElement.Configuration.LinkConfiguration.Display.Automatic ->
         PaymentSheet.LinkConfiguration.Display.Automatic
-    CheckoutController.Configuration.LinkConfiguration.Display.Never ->
+    PaymentElement.Configuration.LinkConfiguration.Display.Never ->
         PaymentSheet.LinkConfiguration.Display.Never
-    CheckoutController.Configuration.LinkConfiguration.Display.WalletButtonHidden ->
+    PaymentElement.Configuration.LinkConfiguration.Display.WalletButtonHidden ->
         PaymentSheet.LinkConfiguration.Display.WalletButtonHidden
 }
